# services/notification/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a notification service API",
        "title": "Notification Service API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost",
    "basePath": "/api/v1",
    "paths": {
        "/notifications": {
            "get": {
                "description": "Returns paginated notification history for a user using cursor-based pagination",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "notifications"
                ],
                "summary": "List notification history",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID (required)",
                        "name": "user_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by status (pending|processing|sent|delivering|failed)",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by channel (email|sms|zalo|push)",
                        "name": "channel",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by event type",
                        "name": "event_type",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Start time RFC3339 (e.g. 2026-01-01T00:00:00Z)",
                        "name": "from",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "End time RFC3339",
                        "name": "to",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Opaque pagination cursor from previous response",
                        "name": "cursor",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page size (default 20, max 100)",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.StandardResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_tojinguyen_notification_internal_dto.ListNotificationsResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.StandardResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "type": "object",
                                            "properties": {
                                                "code": {
                                                    "type": "integer"
                                                },
                                                "message": {
                                                    "type": "string"
                                                }
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.StandardResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "type": "object",
                                            "properties": {
                                                "code": {
                                                    "type": "integer"
                                                },
                                                "message": {
                                                    "type": "string"
                                                }
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/notifications/admin/campaigns": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Create a mass notification campaign",
                "parameters": [
                    {
                        "description": "Campaign request",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_tojinguyen_notification_internal_dto.CreateCampaignRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.StandardResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_tojinguyen_notification_internal_dto.CreateCampaignResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.StandardResponse"
                        }
                    }
                }
            }
        },
        "/notifications/admin/campaigns/{id}/stats": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Get campaign delivery stats",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Campaign ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.StandardResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_tojinguyen_notification_internal_dto.CampaignStatsResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.StandardResponse"
                        }
                    }
                }
            }
        },
        "/notifications/admin/dlq/messages": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "List Dead Letter Queue messages",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter by status (pending|replayed|ignored)",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page limit",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.StandardResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "properties": {
                                                "messages": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/definitions/github_com_tojinguyen_notification_internal_dto.DLQMessageResponse"
                                                    }
                                                },
                                                "total": {
                                                    "type": "integer"
                                                }
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.StandardResponse"
                        }
                    }
                }
            }
        },
        "/notifications/admin/dlq/replay/{id}": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Replay a Dead Letter Queue message",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Message ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.StandardResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.StandardResponse"
                        }
                    }
                }
            }
        },
...
        "/notifications/send": {
            "post": {
                "description": "Direct trigger to send a notification (test or synchonous bypass of queue)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "notifications"
                ],
                "summary": "Send notification",
                "parameters": [
                    {
                        "description": "Notification Event",
                        "name": "event",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_tojinguyen_notification_internal_dto.SendNotificationRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.StandardResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_tojinguyen_notification_internal_dto.SendNotificationResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.StandardResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "type": "object",
                                            "properties": {
                                                "code": {
                                                    "type": "integer"
                                                },
                                                "message": {
                                                    "type": "string"
                                                }
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.StandardResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "type": "object",
                                            "properties": {
                                                "code": {
                                                    "type": "integer"
                                                },
                                                "message": {
                                                    "type": "string"
                                                }
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_tojinguyen_notification_internal_dto.CampaignStatsResponse": {
            "type": "object",
            "properties": {
                "campaign_id": {
                    "type": "string"
                },
                "dispatched_count": {
                    "type": "integer"
                },
                "failed_count": {
                    "type": "integer"
                },
                "last_dispatched_offset": {
                    "type": "integer"
                },
                "pending_count": {
                    "type": "integer"
                },
                "progress_pct": {
                    "type": "number"
                },
                "sent_count": {
                    "type": "integer"
                },
                "status": {
                    "$ref": "#/definitions/github_com_tojinguyen_notification_internal_domain.CampaignStatus"
                },
                "total_recipients": {
                    "type": "integer"
                }
            }
        },
        "github_com_tojinguyen_notification_internal_dto.CreateCampaignRequest": {
            "type": "object",
            "required": [
                "channel",
                "content",
                "event_type",
                "scheduled_at",
                "title"
            ],
            "properties": {
                "channel": {
                    "$ref": "#/definitions/github_com_tojinguyen_notification_internal_domain.NotificationChannel"
                },
                "content": {
                    "type": "string"
                },
                "event_type": {
                    "$ref": "#/definitions/github_com_tojinguyen_notification_internal_domain.EventType"
                },
                "recipients": {
                    "description": "Optional, only used when TargetAudience == \"specific\"",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_tojinguyen_notification_internal_dto.RecipientInput"
                    }
                },
                "scheduled_at": {
                    "type": "string"
                },
                "subject": {
                    "type": "string"
                },
                "target_audience": {
                    "description": "e.g., \"all_users\", \"specific\"",
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "github_com_tojinguyen_notification_internal_dto.CreateCampaignResponse": {
            "type": "object",
            "properties": {
                "campaign_id": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "scheduled_at": {
                    "type": "string"
                },
                "total_recipients": {
                    "type": "integer"
                }
            }
        },
...
        "github_com_tojinguyen_notification_internal_dto.ListNotificationsResponse": {
            "type": "object",
            "properties": {
                "has_more": {
                    "type": "boolean"
                },
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_tojinguyen_notification_internal_dto.NotificationItem"
                    }
                },
                "next_cursor": {
                    "type": "string"
                }
            }
        },
...
        "response.StandardResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {},
                "meta": {},
                "success": {
                    "type": "boolean"
                }
            }
        }
    }
}